ceoa/ceoa-fixed/pyproject.toml
[tool.poetry]
name = "ceoa"
version = "1.0.0"
description = "Compute & Energy Orchestration API - Aetherion CGI Platform"
authors = ["Aetherion Engineering <engineering@aetherion.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
redis = "^5.0.8"
httpx = "^0.27.2"
anthropic = "^0.39.0"
ortools = "^9.11.0"
boto3 = "^1.35.0"
google-cloud-compute = "^1.20.0"
azure-mgmt-compute = "^33.0.0"
prometheus-client = "^0.21.0"
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
black = "^24.8.0"
ruff = "^0.6.9"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
